initial attempt to break amp creation out
Now supports amp-source (this is what we had before),
amp-local (links amp files from ./amps and ./amps_share),
amp-remote (links amps from a remote maven repo.).

The alfresco:amp sub-generator is now setup to dispatch
to these other sub-generators.
