+++ package.json
@@ -14,17 +14,17 @@
     "node": ">=0.10.0"
   },
   "files": [
-    "generators",
-    "generators/action",
-    "generators/app",
-    "generators/amp",
-    "generators/model",
-    "generators/webscript"
+    "generators"
   ],
   "keywords": [
+    "alfresco",
+    "all-in-one",
+    "generator",
+    "yeoman",
     "yeoman-generator"
   ],
   "dependencies": {
+    "adm-zip": "0.4.7",
     "chalk": "1.1.3",
     "cross-spawn": "2.2.3",
     "deasync": "0.1.4",
@@ -34,8 +34,8 @@
     "rmdir": "1.2.0",
     "semver": "5.1.0",
     "split": "1.0.0",
-    "window-size": "^0.2.0",
-    "wrap-ansi": "^2.0.0",
+    "window-size": "0.2.0",
+    "wrap-ansi": "2.0.0",
     "xmldom": "0.1.22",
     "xpath": "0.0.23",
     "yeoman-generator": "0.22.5",
